refactor metadata, add fill value tests
