+++ src/zarr/codecs/_v2.py
@@ -6,9 +6,9 @@ import numcodecs
 from numcodecs.compat import ensure_bytes, ensure_ndarray
 
 from zarr.abc.codec import ArrayArrayCodec, ArrayBytesCodec
-from zarr.array_spec import ArraySpec
 from zarr.buffer import Buffer, NDBuffer
 from zarr.common import JSON, to_thread
+from zarr.metadata.common import ArraySpec
 
 
 @dataclass(frozen=True)
