+++ src/zarr/codecs/pipeline.py
@@ -19,16 +19,16 @@ from zarr.abc.codec import (
 )
 from zarr.abc.store import ByteGetter, ByteSetter
 from zarr.buffer import Buffer, BufferPrototype, NDBuffer
-from zarr.chunk_grids import ChunkGrid
 from zarr.codecs.registry import get_codec_class
 from zarr.common import JSON, ChunkCoords, concurrent_map, parse_named_configuration
 from zarr.config import config
 from zarr.indexing import SelectorTuple, is_scalar, is_total_slice
+from zarr.metadata.common import ChunkGrid
 
 if TYPE_CHECKING:
     from typing_extensions import Self
 
-    from zarr.array_spec import ArraySpec
+    from zarr.metadata.common import ArraySpec
 
 T = TypeVar("T")
 U = TypeVar("U")
