+++ src/zarr/codecs/zstd.py
@@ -7,10 +7,10 @@ import numpy.typing as npt
 from zstandard import ZstdCompressor, ZstdDecompressor
 
 from zarr.abc.codec import BytesBytesCodec
-from zarr.array_spec import ArraySpec
 from zarr.buffer import Buffer, as_numpy_array_wrapper
 from zarr.codecs.registry import register_codec
 from zarr.common import JSON, parse_named_configuration, to_thread
+from zarr.metadata.common import ArraySpec
 
 if TYPE_CHECKING:
     from typing_extensions import Self
